git-svn-id: https://svn.habariproject.org/habari/makaanga/0.x/htdocs/system@3437 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/adminhandler.php
@@ -80,7 +80,7 @@ class AdminHandler extends ActionHandler
 		// Add some default stylesheets
 		Stack::add('admin_stylesheet', array(Site::get_url('admin_theme') . '/css/admin.css', 'screen'), 'admin');
 
-	  	// Add some default template variables
+			// Add some default template variables
 		$this->set_admin_template_vars( $this->theme );
 		$this->theme->admin_type = $type;
 		$this->theme->admin_page = $page;
@@ -94,7 +94,7 @@ class AdminHandler extends ActionHandler
 			Session::error(_t('Access to that page has been denied by the administrator.'));
 			$this->get_blank();
 		}
-
+		
 		switch( $_SERVER['REQUEST_METHOD'] ) {
 			case 'POST':
 				// Let plugins try to handle the page
@@ -138,7 +138,6 @@ class AdminHandler extends ActionHandler
 	 */
 	public function act_admin_ajax()
 	{
-		header( 'Content-Type: application/json' );
 		$context = $this->handler_vars['context'];
 		if ( method_exists( $this, 'ajax_' . $context ) ) {
 			$type = ( isset( $this->handler_vars['content_type'] ) && !empty( $this->handler_vars['content_type'] ) ) ? $this->handler_vars['content_type'] : '';
@@ -392,8 +391,8 @@ class AdminHandler extends ActionHandler
 			'tag_count' => DB::get_value('SELECT count(id) FROM {tags}'),
 			'page_draft_count' => Posts::get( array( 'count' => 1, 'content_type' => Post::type('page'), 'status' => Post::status('draft'), 'user_id' => User::identify()->id ) ),
 			'entry_draft_count' => Posts::get( array( 'count' => 1, 'content_type' => Post::type('entry'), 'status' => Post::status('draft'), 'user_id' => User::identify()->id ) ),
-			'unapproved_comment_count' => Comments::count_total( Comment::STATUS_UNAPPROVED, FALSE ),
-			'spam_comment_count' => Comments::count_total( Comment::STATUS_SPAM, FALSE ),
+			'unapproved_comment_count' => User::identify()->can( 'manage_all_comments' ) ? Comments::count_total( Comment::STATUS_UNAPPROVED, FALSE ) : Comments::count_by_author( User::identify()->id, Comment::STATUS_UNAPPROVED ),
+			'spam_comment_count' => User::identify()->can( 'manage_all_comments' ) ? Comments::count_total( Comment::STATUS_SPAM, FALSE ) : Comments::count_by_author( User::identify()->id, Comment::STATUS_SPAM ),
 			'user_entry_scheduled_count' => Posts::get( array( 'count' => 1, 'content_type' => Post::type( 'any' ), 'status' => Post::status( 'scheduled' ), 'user_id' => User::identify()->id ) ),
 		);
 
@@ -525,7 +524,7 @@ class AdminHandler extends ActionHandler
 		$post->update( $minor );
 
 		$permalink = ( $post->status != Post::status( 'published' ) ) ? $post->permalink . '?preview=1' : $post->permalink;
-		Session::notice( sprintf( _t( 'The post %1$s has been saved as %2$s.' ), sprintf('<a href="%1$s">\'%2$s\'</a>', $permalink, $post->title), Post::status_name( $post->status ) ) );
+		Session::notice( sprintf( _t( 'The post %1$s has been saved as %2$s.' ), sprintf('<a href="%1$s">\'%2$s\'</a>', $permalink, htmlspecialchars( $post->title ) ), Post::status_name( $post->status ) ) );
 		Utils::redirect( URL::get( 'admin', 'page=publish&id=' . $post->id ) );
 	}
 
@@ -588,7 +587,7 @@ class AdminHandler extends ActionHandler
 		}
 		$post = Post::get( array( 'id' => $id, 'status' => Post::status( 'any' ) ) );
 		$post->delete();
-		Session::notice( sprintf( _t( 'Deleted the %1$s titled "%2$s".' ), Post::type_name( $post->content_type ), $post->title ) );
+		Session::notice( sprintf( _t( 'Deleted the %1$s titled "%2$s".' ), Post::type_name( $post->content_type ), htmlspecialchars( $post->title ) ) );
 		Utils::redirect( URL::get( 'admin', 'page=posts&type=' . Post::status( 'any' ) ) );
 	}
 
@@ -677,7 +676,11 @@ class AdminHandler extends ActionHandler
 		$password2->class[] = 'item clear';
 		$password2->type = 'password';
 		$password2->value = '';
-		$password2->add_validator('validate_same', $password1, _t('Passwords must match.'));
+		$delete = $this->handler_vars->filter_keys('delete');
+		// don't validate password match if action is delete
+		if (!isset($delete['delete'])) {
+			$password2->add_validator('validate_same', $password1, _t('Passwords must match.'));
+		}
 
 		// Regional settings
 		$timezones = DateTimeZone::listIdentifiers();
@@ -933,7 +936,9 @@ class AdminHandler extends ActionHandler
 
 	public function get_users()
 	{
-		return $this->post_users();
+		$this->fetch_users();
+
+		$this->theme->display('users');
 	}
 
 	/**
@@ -1424,12 +1429,14 @@ class AdminHandler extends ActionHandler
 
 					case 'delete':
 						// This comment was marked for deletion
+						$to_update = $this->comment_access_filter( $to_update, 'delete' );
 						Comments::delete_these( $to_update );
 						$modstatus['Deleted %d comments'] = count( $to_update );
 						break;
 
 					case 'spam':
-							// This comment was marked as spam
+						// This comment was marked as spam
+						$to_update = $this->comment_access_filter( $to_update, 'edit' );
 						Comments::moderate_these( $to_update, Comment::STATUS_SPAM );
 						$modstatus['Marked %d comments as spam'] = count( $to_update );
 						break;
@@ -1437,6 +1444,7 @@ class AdminHandler extends ActionHandler
 					case 'approve':
 					case 'approved':
 						// Comments marked for approval
+						$to_update = $this->comment_access_filter( $to_update, 'edit' );
 						Comments::moderate_these( $to_update, Comment::STATUS_APPROVED );
 						$modstatus['Approved %d comments'] = count( $to_update );
 						foreach ( $to_update as $comment ) {
@@ -1447,11 +1455,13 @@ class AdminHandler extends ActionHandler
 					case 'unapprove':
 					case 'unapproved':
 						// This comment was marked for unapproval
+						$to_update = $this->comment_access_filter( $to_update, 'edit' );
 						Comments::moderate_these( $to_update, Comment::STATUS_UNAPPROVED );
 						$modstatus['Unapproved %d comments'] = count ( $to_update );
 						break;
 
 					case 'edit':
+						$to_update = $this->comment_access_filter( $to_update, 'edit' );
 						foreach ( $to_update as $comment ) {
 							// This comment was edited
 							if ( $_POST['name_' . $comment->id] != NULL ) {
@@ -1466,6 +1476,8 @@ class AdminHandler extends ActionHandler
 							if ( $_POST['content_' . $comment->id] != NULL ) {
 								$comment->content = $_POST['content_' . $comment->id];
 							}
+
+							$comment->update();
 						}
 						$modstatus['Edited %d comments'] = count( $to_update );
 						break;
@@ -1495,6 +1507,11 @@ class AdminHandler extends ActionHandler
 			'offset' => $offset,
 			'orderby' => $orderby,
 		);
+
+		// only get comments the user is allowed to manage
+		if( !User::identify()->can( 'manage_all_comments' ) ) {
+			$arguments['post_author'] = User::identify()->id;
+		}
 
 		// there is no explicit 'all' type/status for comments, so we need to unset these arguments
 		// if that's what we want. At the same time we can set up the search field
@@ -1534,25 +1551,51 @@ class AdminHandler extends ActionHandler
 		$baseactions = array();
 		$statuses = Comment::list_comment_statuses();
 		foreach ( $statuses as $statusid => $statusname ) {
-			$baseactions[$statusname] = array('url' => 'javascript:itemManage.update(\'' . $statusname . '\',__commentid__);', 'title' => _t('Change this comment\'s status to %s', array($statusname)), 'label' => Comment::status_action($statusid));
+			$baseactions[$statusname] = array('url' => 'javascript:itemManage.update(\'' . $statusname . '\',__commentid__);', 'title' => _t('Change this comment\'s status to %s', array($statusname)), 'label' => Comment::status_action($statusid), 'access' => 'edit' );
 		}
 
 		/* Standard actions */
-		$baseactions['delete'] = array('url' => 'javascript:itemManage.update(\'delete\',__commentid__);', 'title' => _t('Delete this comment'), 'label' => _t('Delete'));
-		$baseactions['edit'] = array('url' => URL::get('admin', 'page=comment&id=__commentid__'), 'title' => _t('Edit this comment'), 'label' => _t('Edit'));
+		$baseactions['delete'] = array('url' => 'javascript:itemManage.update(\'delete\',__commentid__);', 'title' => _t('Delete this comment'), 'label' => _t('Delete'), 'access' => 'delete' );
+		$baseactions['edit'] = array('url' => URL::get('admin', 'page=comment&id=__commentid__'), 'title' => _t('Edit this comment'), 'label' => _t('Edit'), 'access' => 'edit' );
 
 		/* Actions for inline edit */
-		$baseactions['submit'] = array('url' => 'javascript:inEdit.update();', 'title' => _t('Submit changes'), 'label' => _t('Update'), 'nodisplay' => TRUE);
+		$baseactions['submit'] = array('url' => 'javascript:inEdit.update();', 'title' => _t('Submit changes'), 'label' => _t('Update'), 'nodisplay' => TRUE, 'access' => 'edit' );
 		$baseactions['cancel'] = array('url' => 'javascript:inEdit.deactivate();', 'title' => _t('Cancel changes'), 'label' => _t('Cancel'), 'nodisplay' => TRUE);
 
 		/* Allow plugins to apply actions */
 		$actions = Plugins::filter('comments_actions', $baseactions, $this->theme->comments);
 
 		foreach ( $this->theme->comments as $comment ) {
-			$menu = $actions;
+			// filter the actions based on the user's permissions
+			$comment_access = $comment->get_access();
+			$menu = array();
+			foreach ( $actions as $name => $action ) {
+				if ( !isset( $action['access'] ) || ACL::access_check( $comment_access, $action['access'] ) ) {
+					$menu[$name] = $action;
+				}
+			}
+			// remove the current status from the dropmenu
 			unset($menu[Comment::status_name($comment->status)]);
 			$comment->menu = Plugins::filter('comment_actions', $menu, $comment);
 		}
+	}
+
+	/**
+	 * A helper function for fetch_comments()
+	 * Filters a list of comments by ACL access
+	 * @param object $comments an array of Comment objects
+	 * @param string $access the access type to check for
+	 * @return a filtered array of Comment objects.
+	 */
+	public function comment_access_filter( $comments, $access )
+	{
+		$result = array();
+		foreach ( $comments as $comment ) {
+			if ( ACL::access_check( $comment->get_access(), $access ) ) {
+				$result[] = $comment;
+			}
+		}
+		return $result;
 	}
 
 	/**
@@ -1718,14 +1761,20 @@ class AdminHandler extends ActionHandler
 				foreach ( $to_update as $post ) {
 					switch( $change ) {
 					case 'delete':
-						$post->delete();
+						if ( ACL::access_check( $post->get_access(), 'delete' ) ) {
+							$post->delete();
+						}
 						break;
 					case 'publish':
-						$post->publish();
+						if ( ACL::access_check( $post->get_access(), 'edit') ) {
+							$post->publish();
+						}
 						break;
 					case 'unpublish':
-						$post->status = Post::status( 'draft' );
-						$post->update();
+						if ( ACL::access_check( $post->get_access(), 'edit') ) {
+							$post->status = Post::status( 'draft' );
+							$post->update();
+						}
 						break;
 					}
 				}
@@ -1888,7 +1937,9 @@ class AdminHandler extends ActionHandler
 		$item_ids = array();
 
 		foreach ( $this->theme->posts as $post ) {
+			if( ACL::access_check($post->get_access(), 'delete' ) ) {
 			$item_ids['p' . $post->id] = 1;
+		}
 		}
 
 		$output = array(
@@ -1965,6 +2016,11 @@ class AdminHandler extends ActionHandler
 		}
 
 		$comment = Comment::get($handler_vars['id']);
+		if ( !ACL::access_check( $comment->get_access(), 'edit' ) ) {
+			Session::error( _t('You do not have permission to edit this comment.') );
+			echo Session::messages_get( true, array( 'Format', 'json_messages' ) );
+			return;
+		}
 
 		if ( isset($handler_vars['author']) && $handler_vars['author'] != '' ) {
 			$comment->name = $handler_vars['author'];
@@ -2013,11 +2069,19 @@ class AdminHandler extends ActionHandler
 			}
 		}
 		$posts = Posts::get( array( 'id' => $ids, 'nolimit' => true ) );
+		$deleted = 0;
 		foreach ( $posts as $post ) {
+			if ( ACL::access_check( $post->get_access(), 'delete' ) ) {
 			$post->delete();
+				$deleted++;
+		}
 		}
 
-		Session::notice( sprintf( _t('Deleted %d entries.'), count($posts) ) );
+		Session::notice( sprintf( _t('Deleted %d entries.'), $deleted ) );
+		if( $deleted != count( $posts ) ) {
+			Session::notice( _t( 'You did not have permission to delete some entries.' ) );
+		}
+
 		echo Session::messages_get( true, array( 'Format', 'json_messages' ) );
 	}
 
@@ -2493,13 +2557,16 @@ class AdminHandler extends ActionHandler
 
 	public function get_groups()
 	{
-		$this->post_groups();
+		// prepare the WSSE tokens
+		$this->theme->wsse = Utils::WSSE();
+
+		$this->theme->groups = UserGroups::get_all();
+
+		$this->display( 'groups' );
 	}
 
 	public function post_groups()
 	{
-		Utils::check_request_method( array( 'POST' ) );
-
 		// prepare the WSSE tokens
 		$this->theme->wsse = Utils::WSSE();
 
@@ -2679,7 +2746,6 @@ class AdminHandler extends ActionHandler
 				Session::notice( $msg_status );
 				echo Session::messages_get( true, array( 'Format', 'json_messages' ) );
 				break;
-		Utils::check_request_method( array( 'POST' ) );
 
 			case 'rename':
 				if ( isset($this->handler_vars['master']) ) {
@@ -2719,8 +2785,6 @@ class AdminHandler extends ActionHandler
 	*/
 	public function get_sysinfo()
 	{
-		Utils::check_request_method( array( 'POST' ) );
-
 		$sysinfo = array();
 		$siteinfo = array();
 
@@ -2969,8 +3033,6 @@ class AdminHandler extends ActionHandler
 				$require_any = array( 'manage_import' => true );
 				break;
 			case 'users':
-		Utils::check_request_method( array( 'POST' ) );
-
 			case 'user':
 			case 'ajax_update_users':
 			case 'ajax_users':
@@ -3018,7 +3080,7 @@ class AdminHandler extends ActionHandler
 				break;
 			case 'sysinfo': 
 				$require_any = array( 'super_user' => true ); 
-				break; 
+				break;
 			case 'dashboard':
 			case 'ajax_dashboard':
 				$result = true;
@@ -3117,6 +3179,7 @@ class AdminHandler extends ActionHandler
 		}
 		$output['controls'] = $controls_out;
 
+		header( 'content-type:text/javascript' ); 
 		echo json_encode( $output );
 	}
 
@@ -3131,8 +3194,8 @@ class AdminHandler extends ActionHandler
 
 		$panel = '';
 		$panel = Plugins::filter( 'media_panels', $panel, $silo, $rpath, $panelname );
-		$rootpath = strpos($path, '/') !== false ? substr($path, 0, strpos($path, '/')) : $path;
-		$controls = array('root' => '<a href="#" onclick="habari.media.fullReload();habari.media.showdir(\''. $rootpath . '\');return false;">' . _t('Root') . '</a>');
+
+		$controls = array();
 		$controls = Plugins::filter( 'media_controls', $controls, $silo, $rpath, $panelname );
 		$controls_out = '';
 		foreach ( $controls as $k => $v ) {
@@ -3201,7 +3264,6 @@ class AdminHandler extends ActionHandler
 		Stack::add( 'admin_header_javascript', Site::get_url('scripts') . "/ui.tabs.js", 'ui.tabs', array('jquery', 'ui.core') );
 		Stack::add( 'admin_header_javascript', Site::get_url('scripts') . "/ui.sortable.js", 'ui.sortable', array('jquery', 'ui.core') );
 		Stack::add( 'admin_header_javascript', Site::get_url('scripts') . "/ui.resizable.js", 'ui.resizable', array('jquery', 'ui.core') );
-		Stack::add( 'admin_header_javascript', Site::get_url('scripts') . "/jquery.spinner.js", 'jquery.spinner', 'jquery' );
 		Stack::add( 'admin_header_javascript', Site::get_url('scripts') . "/jquery.color.js", 'jquery.color', 'jquery' );
 		Stack::add( 'admin_header_javascript', Site::get_url('habari') . "/3rdparty/humanmsg/humanmsg.js", 'humanmsg', 'jquery' );
 		Stack::add( 'admin_header_javascript', Site::get_url('habari') . "/3rdparty/hotkeys/jquery.hotkeys.js", 'jquery.hotkeys', 'jquery' );
